ML-Based-Test-Automation/requirements.txt
# Core Libraries
numpy==1.22.4
pandas==1.4.3
scikit-learn==1.1.2
scipy==1.8.1
matplotlib==3.5.2
seaborn==0.11.2

# Machine Learning
tensorflow==2.9.1
xgboost==1.6.1
lightgbm==3.3.2
shap==0.41.0
eli5==0.13.0
imbalanced-learn==0.9.1

# Data Processing
pytest==7.1.2
pytest-xdist==2.5.0
pylint==2.14.5
black==22.6.0
flake8==5.0.4

# Jupyter
jupyter==1.0.0
notebook==6.4.12
ipykernel==6.15.1

# Visualization
plotly==5.9.0
dash==2.6.1

# CI/CD Integration
pytest-html==3.1.1
pytest-cov==3.0.0
pytest-mock==3.8.2

# Utilities
python-dotenv==0.20.0
tqdm==4.64.0
click==8.1.3
joblib==1.1.0
requests==2.28.1
pyyaml==6.0
daiquiri==3.0.1 